Fix circular imports and update tests to work with async functionality
